# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The API documentation for the golang backend server.",
        "title": "Go Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/meta": {
            "get": {
                "description": "Retrieves metadata associated with a given UUID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meta"
                ],
                "summary": "Get metadata by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The UUID of the metadata to retrieve",
                        "name": "documentUUID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful retrieval of metadata",
                        "schema": {
                            "$ref": "#/definitions/models.Meta"
                        }
                    },
                    "400": {
                        "description": "Bad request, typically due to missing/invalid UUID"
                    },
                    "500": {
                        "description": "Internal server error, "
                    }
                }
            },
            "put": {
                "description": "Updates specific fields of an existing metadata entry.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meta"
                ],
                "summary": "Update existing metadata",
                "parameters": [
                    {
                        "description": "Metadata update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.UpdateMetaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful update"
                    },
                    "400": {
                        "description": "Bad request, typically due to invalid input"
                    },
                    "500": {
                        "description": "Internal server error, "
                    }
                }
            },
            "post": {
                "description": "Creates new metadata with a generated UUID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meta"
                ],
                "summary": "Add new metadata",
                "parameters": [
                    {
                        "description": "Metadata creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.AddMetaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful creation, returns the metadata UUID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, typically due to invalid input"
                    },
                    "500": {
                        "description": "Internal server error, "
                    }
                }
            },
            "delete": {
                "description": "Deletes metadata based on the provided UUID in the request body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meta"
                ],
                "summary": "Delete metadata by UUID",
                "parameters": [
                    {
                        "description": "Metadata deletion request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.DeleteMetaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful deletion",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, typically due to invalid input"
                    },
                    "500": {
                        "description": "Internal server error, "
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Meta": {
            "type": "object",
            "properties": {
                "documentUUID": {
                    "type": "string",
                    "example": "ba3ca973-5052-4030-a528-39b49736d8ad"
                },
                "height": {
                    "type": "number",
                    "example": 1080
                },
                "images": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "numberOfPages": {
                    "type": "integer",
                    "example": 31
                },
                "ownerType": {
                    "type": "string",
                    "example": "1"
                },
                "ownerUUID": {
                    "type": "string",
                    "example": "34906041-2d68-45a2-9671-9f0ba89f31a9"
                },
                "width": {
                    "type": "number",
                    "example": 1920
                }
            }
        },
...
        "v1.AddMetaRequest": {
            "type": "object",
            "properties": {
                "height": {
                    "type": "number",
                    "format": "float32"
                },
                "images": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "numberOfPages": {
                    "type": "integer",
                    "format": "int32"
                },
                "width": {
                    "type": "number",
                    "format": "float32"
                }
            }
        },
...
        "v1.DeleteMetaRequest": {
            "type": "object",
            "properties": {
                "uuid": {
                    "type": "string"
                }
            }
        },
        "v1.UpdateMetaRequest": {
            "type": "object",
            "properties": {
                "height": {
                    "type": "number",
                    "format": "float32"
                },
                "images": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "numberOfPages": {
                    "type": "integer",
                    "format": "int32"
                },
                "uuid": {
                    "type": "string"
                },
                "width": {
                    "type": "number",
                    "format": "float32"
                }
            }
        }
    },
...
}